Premier commit - Projet Video Center

+++ src/Controller/VideoController.php
@@ -25,11 +25,18 @@ class VideoController extends AbstractController
     #[Route('/video/create', name: 'app_video_create')]
     public function create(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->getUser()){
+            $this->addFlash('danger', 'Vous devez être connecté pour créer une vidéo.');
+            return $this->redirectToRoute('app_login');
+        }
+
         $video = new Video();
         $form = $this->createForm(VideoType::class, $video);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            $video->setUser($this->getUser());
             $entityManager->persist($video);
             $entityManager->flush();
 
@@ -53,23 +60,53 @@ class VideoController extends AbstractController
 #[Route('/video/{id}/edit', name: 'app_video_edit')]
 public function edit(Request $request, Video $video, EntityManagerInterface $em): Response
 {
+    if (!$this->getUser()){
+        $this->addFlash('danger', 'Vous devez être connecté pour modifier une vidéo.');
+        return $this->redirectToRoute('app_login');
+    }
+
+
+    // vérifier si l'utilsateur est bien le propriétaire de la vidéo
+    if($video->getUser() !== $this->getUser()){
+        $this->addFlash('danger', 'Vous n\'êtes pas autorisé.');
+        return $this->redirectToRoute('app_login');
+    }
+    // si on arrive ici l'utilisatteur est bien le vrai utisateur
+
     $form = $this->createForm(VideoType::class, $video);
     $form->handleRequest($request);
     if ($form->isSubmitted() && $form->isValid()){
         $em->flush();
-        $this->addFlash('sucess', 'vidéo modifié !');
+        $this->addFlash('success', 'vidéo modifié !');
         return $this->redirectToRoute('app_home');
     }
-    return $this->render('video/edit.html.twig', ['form' => $form->createView(), 'video' => $video]);
+
+    return $this->render('video/edit.html.twig', [
+        'form' => $form->createView(), 
+        'video' => $video
+    ]);
 }
 
 #[Route('/video/{id}/delete', name: 'app_video_delete')]
 public function delete(Request $request, Video $video, EntityManagerInterface $em): Response
 {
+
+    // vérifier si l'utilsateur est connecté
+        if (!$this->getUser()){
+        $this->addFlash('danger', 'Vous devez être connecté pour modifier une vidéo.');
+        return $this->redirectToRoute('app_login');
+    }
+    // si on arrive ici l'utilisatteur est bien le vrai utisateur
+    if($video->getUser() !== $this->getUser()){
+        $this->addFlash('danger', 'Vous n\'êtes pas autorisé.');
+        return $this->redirectToRoute('app_login');
+    }
+
     if ($this->isCsrfTokenValid('delete' . $video->getId(), $request->request->get('_token'))) {
+        
         $em->remove($video);
         $em->flush();
-        $this->addFlash('sucess', 'Vidéo supprimée !');
+        $this->addFlash('success', 'Vidéo supprimée !');
 
     }
     return $this->redirectToRoute('app_home');
